[client] [dashboard] [accounts_head] [new_account_form] :: Combining modal also with the previously created component
